Much closer to full implementation of bubble sort
Added the full function and have it sorting the array, only problem is it is calling the update function of the slider in order to get the returned SLIDER_DATA array. This means it is correctly sorting a new array that is different than the one on the screen before the button is pressed. Not sure how to get around this just yet, might make the SLIDER_DATA variable global so that it is easier to work with (bad practice I know)