Got bubblesort to work!
After fixing the problem of trying to use d3.js to graph an object instead of an array of objects, it worked! Only problem was that I was graphing the next randomly generated array due to the update() function being called inside of the bubblesort function. Moving the line of code that assigns this update() to the SLIDER_DATA array outside of the bubblesort function allows me to display the sorted version of the graph currently being viewed by the user. Last step is to try and animate each step in the sorting so that the functions can be fully visualized.