Added new function
Began adding a new function to handle displaying the sorted array from the bubblesort function. Hopefully this function will be able to handle the visualization of the graph, such as moving and swapping the bars, instead of doing it all in one function. I plan to have this function get called from the other function once certain values are swapped, and then do checks within the second function to determine if they should be colored red or green.