Removed bubblearr
In order to uncomplicate things temporarily, I replaced the vals array with the SLIDER_DATA array and am swapping those values directly now instead of swapping them in the vals array and then creating a new array of objects with bubblearr. Hopefully this should allow me to better detect where the sorting is going wrong.