Taking a different approach to bubblesort
Deleted the copied code and am now trying an alternative approach of combining elements from my original code and the copied code to try and get the visualization to work. I figured this would be easier than trying to work with both versions of the code simultaneously to get them to match.