Refactoring
After asking for help on reddit, I was informed I was using an old way of join data in d3. Instead of enter() and exit() I can now just use .join() which is easier. I also am going to keep trying to implement a key function which will keep my data and elements bound together even when they are added, removed, or changed.