Changed some files
Got the original file back to how it was and am editing it now while also messing around with the new file until I can achieve the desired result on either of them. Made tiny progress in the new file; slowly making it more modular to zone in on what the problem is. May have something to do with my bubblesort function