Started creation of a test case
Decided my best option to get unstuck is to create a test case that focuses just on visualizing the sorting process. Therefore, these new files/case will not have a generate array button, a slider, or any other button besides the sorting button. Additionally, I plan to simplify the data stream (an array of integers instead of an array of objects with values and ids like before) to hopefully make things even simpler. If I can manage to get this case to work, I may modify it or the original files in order to get the full program working how I want. Fingers crossed