Made some progress
I've got the draw function working to remove the graph after each call to it is made. Bubble sort is working in tandem with this to draw the sorted graph. Just have to now figure out the original problem of not being able to show each step of the sorting. I'm thinking using setInterval or chunks might be the answer