Made a little more progress today
Decided to try and implement insertion sort as it was less complicated and required no recursion like quicksort. I was able to get it working fully like the others, and in doing so I discovered that is I await and draw inside the for loop that populates bubblearr, it not only shows every step of the sorting, but also works with the draw function more fully and turns the currently selected node green just like bubblesort does. This proved true for selection sort too, however generating a new array after sorting proved to break the program a bit which will have to be looked into.