Pretty much fixed the bug
Realized I needed to be emptying the ids array alongside the vals array. This is why the bubblesort array was growing everytime the array was generated, as the vals were being erased but the ids were not, which is why the values were undefined. Adding this in kept the array size the same as DUMMY_DATA.
I am now trying to switch array from the global variable way of doing things, trying to create unique ids, vals, and potentially bubblearrs for each sorting algorithm.