Made some progress on quicksort
Got it to work and draw (something) using the async function, just not sure what. I know that because of the way quicksort works, it is splitting the main array into smaller and smaller arrays, which is messing with the drawing mechanism. May have to settle on having it show the final sorted graph, although I am going to try and show each step. 
Also working on resolving the undefined issue when adding the quickids parameter to the functions.